shared/simplestreams: Split index/manifest out
